% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pop_ut_tract}
\alias{df_pop_ut_tract}
\title{A data frame containing population estimates for Utah Census Tracts in 2013.}
\usage{
data(df_pop_ut_tract)
}
\description{
A data frame containing population estimates for Utah Census Tracts in 2013.
}
\examples{
data(df_pop_ut_tract)

# zoom in on Salt Lake City, which has FIPS code 49035
ut_tract_choropleth(df_pop_ut_tract, 
                    title  = "2013 Salt Lake County Census Tract\\n Population Estimates", 
                    legend = "Population",
                    county_zoom = 49035)
\dontrun{
ut_tract_choropleth(df_pop_ut_tract, 
                    title  = "2013 Utah Census Tract\\n Population Estimates", 
                    legend = "Population")

}
}
\references{
Taken from the US American Community Survey (ACS) 5 year estimates.
}
\seealso{
\link{df_ut_tract_demographics}
}

