% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_ut_tract_demographics}
\alias{df_ut_tract_demographics}
\title{A data frame containing demographic statistics for each Census Tract in Utah.}
\usage{
data(df_ut_tract_demographics)
}
\description{
A data frame containing demographic statistics for each Census Tract in Utah.
}
\examples{
 # 2013 per capita income estimate
 data(df_ut_tract_demographics)
 colnames(df_ut_tract_demographics)
 
 df_ut_tract_demographics$value = df_ut_tract_demographics$per_capita
 ut_tract_choropleth(df_ut_tract_demographics,
                     title       = "2013 Salt Lake Census Tract\\n Per Capita Income",
                     legend      = "Income",
                     num_colors  = 1,
                     county_zoom = 49035)
}
\references{
Data comes from the 2013 5-year American Community Survey (ACS). Data generated by
?get_ut_tract_demographics.
}
\keyword{data}

