% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ut_tract_demographics.R
\name{get_ut_tract_demographics}
\alias{get_ut_tract_demographics}
\title{Get a handful of demographic variables on Utah Census Tracts from the US Census Bureau as a data.frame.}
\usage{
get_ut_tract_demographics(endyear = 2014, span = 5)
}
\arguments{
\item{endyear}{The end year for the survey}

\item{span}{The span of the survey}
}
\description{
The data comes from the American Community Survey (ACS). The variables are: total population, percent White 
not Hispanic, Percent Black or African American not Hispanic, percent Asian not Hispanic,
percent Hispanic all races, per-capita income, median rent and median age.
}
\examples{
\dontrun{
df = get_ut_tract_demographics(endyear=2010, span=5)
colnames(df)

# analyze the percent of people who are white not hispanic
# a boxplot shows the distribution
boxplot(df$percent_white)

# a choropleth map shows the location of the values in utah
# set the 'value' column to be the column we want to render
df$value = df$percent_white
ut_tract_choropleth(df, 
                    title="2010 Census Tracts\\nPercent White not Hispanic", 
                    legend="Percent")

# zoom into salt lake county
ut_tract_choropleth(df, 
                    title="2010 Census Tracts\\nPercent White not Hispanic", 
                    legend="Percent",
                    county_zoom=49035)
}
}
\references{
The choroplethr guide to Census data: http://cran.r-project.org/web/packages/choroplethr/vignettes/e-mapping-us-census-data.html
}

