% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ut.tract.regions}
\alias{ut.tract.regions}
\title{Metadata about the Utah Census Tracts map}
\usage{
data(ut.tract.regions)
}
\description{
In addition to the region/GEOID it contains the TRACTCE and county FIPS code
as both an integer and character
}
\seealso{
\link{ut.tract.map}
}

