% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut_tract_choropleth.R
\name{ut_tract_choropleth}
\alias{ut_tract_choropleth}
\title{Create a choropleth of US Census Tracts in Utah}
\usage{
ut_tract_choropleth(df, title = "", legend = "", num_colors = 7,
  tract_zoom = NULL, county_zoom = NULL, reference_map = FALSE)
}
\arguments{
\item{df}{A data frame with a column named "region" and a column named "value".  Elements in 
the "region" column must exactly match how census tracts are labelled in the "region" column in ?ut.tract.regions}

\item{title}{An optional title for the map.}

\item{legend}{An optional name for the legend.}

\item{num_colors}{The number of colors on the map. A value of 1 
will use a continuous scale. A value in [2, 9] will use that many colors.}

\item{tract_zoom}{An optional vector of tracts to zoom in on. Elements of this vector must exactly 
match the names of tracts as they appear in the "region" column of ?ut.tract.regions.}

\item{county_zoom}{An optional vector of county FIPS codes to zoom in on. Elements of this 
vector must exactly match the names of counties as they appear in the "county.fips.numeric" column 
of ?ut.tract.regions.}

\item{reference_map}{If true, render the choropleth over a reference map from Google Maps.}
}
\description{
Create a choropleth of US Census Tracts in Utah
}
\examples{
# zoom in on Salt Lake, which has FIPS code 49035
data(df_pop_ut_tract)
ut_tract_choropleth(df_pop_ut_tract,
                    title  = "2013 Salt Lake Census Tract\\n Population Estimates",
                    legend = "Population",
                    county_zoom = 49035)                  

# add a reference map
ut_tract_choropleth(df_pop_ut_tract,
                    title  = "2013 Salt Lake Census Tract\\n Population Estimates",
                    legend        = "Population",
                    county_zoom   = 49035,
                    reference_map = TRUE)                  

\dontrun{

ut_tract_choropleth(df_pop_ut_tract,
                    title  = "2013 Utah Census Tract\\n Population Estimates",
                    legend = "Population")

# 2013 per capita income estimate
data(df_ut_tract_demographics)
df_ut_tract_demographics$value = df_ut_tract_demographics$per_capita
ut_tract_choropleth(df_ut_tract_demographics,
                    title         = "2013 Salt Lake Census Tract\\n Per Capita Income",
                    legend        = "Income",
                    num_colors    = 4,
                    county_zoom   = 49035,
                    reference_map = TRUE)
}
}
\seealso{
\url{https://www.census.gov/geo/reference/gtc/gtc_ct.html} for more information on Census Tracts
}

