% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut_tract_choropleth.R
\docType{data}
\name{UtTractChoropleth}
\alias{UtTractChoropleth}
\title{An R6 object for creating choropleths of Utah Census Tracts.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UtTractChoropleth
}
\description{
An R6 object for creating choropleths of Utah Census Tracts.
}
\keyword{datasets}

